fixing usage issue
