--- example/translations/en.json
@@ -1,5 +1,0 @@
-{
-    "greeting": "Hello, World!",
-    "welcome": "Welcome to our application.",
-    "home.title": "Home"
-}
