--- example/translations/fr.json
@@ -1,5 +1,0 @@
-{
-    "greeting": "Bonjour le monde!",
-    "welcome": "Bienvenue dans notre application.",
-    "home.title": "Maison"
-}
