--- src/index.js
@@ -1,3 +1,0 @@
-import { LocalizationProvider, useTranslation } from './Localization';
-
-export default { LocalizationProvider, useTranslation };
